ingestion_plugin_duc/requirements.txt
dify_plugin>=0.4.0,<0.7.0
psycopg2-binary
pandas
requests
